Flytt metoder for å oppdatere state ett hakk inn
